+++ src/Hacks/NoFlip.cpp
@@ -9,6 +9,9 @@ class $modify (GJBaseGameLayer)
 {
     void toggleFlipped(bool p0, bool p1)
     {
+        if (Client::GetModuleEnabled("instant-reverse"))
+            p1 = true;
+
         if (!Client::GetModuleEnabled("no-reverse"))
             GJBaseGameLayer::toggleFlipped(p0, p1);
     }