+++ src/Keybinds/RecordKeyStruct.cpp
@@ -1,6 +1,6 @@
 #include "RecordKeyStruct.hpp"
 #include <Geode/ui/GeodeUI.hpp>
-#include "../Utils/CCBlurLayer.hpp"
+#include "../UI/BlurLayer.hpp"
 #include "../Client/Client.h"
 
 bool RecordKeyStruct::init(geode::utils::MiniFunction<void(KeyStruct)> obj)
@@ -16,7 +16,7 @@ bool RecordKeyStruct::init(geode::utils::MiniFunction<void(KeyStruct)> obj)
 
     if (Client::GetModuleEnabled("menu-bg-blur"))
     {
-        auto blur = CCBlurLayer::create();
+        auto blur = BlurLayer::create();
         blur->runAction(CCEaseIn::create(CCFadeTo::create(0.5f, 255), 2));
         this->addChild(blur);
     }